A suggestion from -- wait, the envelope please -- And what a surprise we have for you tonight, ladies and gentlemen!  This one comes from Greg Stein, of Palo Alto, California!  Welcome to the show, Greg, glad to have you with us.
   "Thanks, glad to be here."

So have you been programming Python long, Greg?

   "Guido invented Python; but I invented Guido."

Heh-heh.  Well, let's cut right to the chase, Greg.  If you had to
compute the length of number expressed in decimal, would you import
the math module and take a logarithm, or would you just format the
number as a string and take the length of the string?

   "Is this a trick question?"

Actually, no:

* tools/cvs2svn/cvs2svn.py: Stop importing math.
  (Dump.start_revision): Don't use math to get the length of the
    length, just format it and ask.