Fix a problem with cvs2svn.py introduced by r8587 that caused lots of extra copies when creating tags and branches. The problem is in the function SymbolicNameTracker.cleanup_entries, which was added in r8587 tries to remove sub-components that are not copied using the optimal revision.
If the parent has an optimal range of revs N-M, N is selected as the
best rev. The child has another optimal range I-J. If the intervals
overlap for revision N, i.e. I < N and J >= N, then N is an optimal
revision for the parent as well as the child. The problem is that
SymbolicNameTracker.cleanup_entries simply compares N and I, which is
too strict. This patch verifies that N is inside the child's optimal
interval using the new function SymbolicNameTracker.is_best_rev.

* tools/cvs2svn/cvs2svn.py:

  (SymbolicNameTracker.is_best_rev): New function.

  (SymbolicNameTracker.cleanup_entries): Add a limit_rev arg.  Use
  is_best_rev when filtering out components.

  (SymbolicNameTracker.copy_descend): Supply dumper.revision to
  cleanup_entries.

* tools/cvs2svn/run-tests.py:

  (simple_tags): Update the rev numbers to match the updated
  cvs2svn.py. It is now the same as before r8587 (modulo another
  unrelated change that has happened since then).