With Fitz, don't try to re-add files that already exist.
The bug this fixes can be reproduced by producing a dumpfile via
DumpfileDelegate (that is, adjust the delegation in pass8, then run
'./cvs2svn.py --dump-only --skip-cleanup test-data/main-cvsrepos/'),
and then trying to load the dumpfile into an empty repository.  The
load will fail in r3, because that revision attempts to add a file
that already exists, because the file was in a copied subtree in r2.

With this change, the load makes it all the way through r34, before
failing in r35 with a suspiciously similar error.

* cvs2svn.py
  (SVNRepositoryMirror.commit): Don't make the change if it's a no-op
    CVSRevision, according to the empty-deltatext/1.1.1.1/etc rules.