Fix issue #7: Support multicomponent --trunk, --tags, and --branches ("TTB").
A whole scad of commits have led up to this one.  This commit
activates the feature by (1) removing the parameter check that
previously prohibited it, and (2) adding a bunch of test cases for
multicomponent TTB directories.

The added tests increase the run time of run-tests.py (on my system)
by about 30% because they require several additional cvs2svn runs.  If
this added time is considered to be a problem, some of the test
variants could undoubtedly be removed--there is a big overkill factor
here, especially considering that all the tests worked the first try.

For now, the TTB directories are still required to be disjoint,
because there doesn't seem to have been a consensus on the mailing
list about the preferred set of rules.  If a consensus emerges, I'll
be happy to rework the code.

* cvs2svn (normalize_ttb_path): Don't insist on TTB directories being
  top-level.

* run-tests.py (prune_with_care, simple_tags, phoenix_branch,
  no_trunk_prune, tagged_branch_and_trunk, branch_delete_first,
  empty_trunk, peer_path_pruning): Added a **kw parameter, allowing
  "trunk", "branches", and "tags" arguments to be specified.  For the
  most part, these values are simply passed through to
  ensure_conversion().

  (prune_with_care_variant, simple_tags_variant,
  phoenix_branch_variant, no_trunk_prune_variant,
  tagged_branch_and_trunk_variant, branch_delete_first_variant,
  empty_trunk_variant, peer_path_pruning_variant): New test functions.
  These call the corresponding non-_variant methods with variant
  values for the TTB options.

  (nested_ttb_directories): Added some tests for multicomponent TTB
  directories.