Fixed problem with r1675 and r1676 (SDatabase didn't work with bsddb3).
bsddb3's connection objects don't have get() methods, so cvs2svn was
failing miserably on my computer after I installed that module.  This
is because the new SDatabase factory function didn't supply the get()
method as is done by Database.  This change moves the common code to
an AbstractDatabase class and derives SDatabase and Database from
that.

* cvs2svn (AbstractDatabase): New abstract class.

  (SDatabase): Converted from a factory function back into a class.

  (Database): Derive from AbstractDatabase.