Improved CVSRepository.get_cvs_path().
Previously, if cvs_repos_path was 'a', this method would happily
convert 'abc,v' to 'bc'.  This change requires an os.sep to follow the
cvs_repos_path (except in the case that the argument is equal to
cvs_repos_path).  It also doesn't insist on the presence of ',v'
anymore--if it is present, it is stripped off, but if not, that is OK
too.

This flexibility will be required when multiproject conversions are
supported.

* cvs2svn (CVSRepository.get_cvs_path): Changed the method to use a
  regular expression to strip off the directory part (including proper
  handling of os.sep).