Clean up handling of STATISTICS_FILE by ArtifactManager.
Change ArtifactManager to keep track of all passes that are currently
being executed (allowing more than one at a time).  Require
pass_started() to be registered, and add a new method pass_continued()
to indicate that the pass will be continued during a future run of the
program.

Change PassManager to register itself as starting at the beginning of
the run, and either done or continued at the end of the run.  It uses
the STATISTICS_FILE and therefore the other passes don't have to
register that they use it.

* cvs2svn_lib/artifact_manager.py (ArtifactManager): Updated class
  docstring.

  (ArtifactManager._active_passes): New member.

  (ArtifactManager.__init__): Initialize _active_passes.

  (ArtifactManager.pass_started, ArtifactManager.pass_continued): New
  methods.

  (ArtifactManager.pass_done): Delete WHICH_PASS from _active_passes.

* cvs2svn_lib/pass_manager.py (PassManager.run): Don't need to
  register each pass depending on STATISTICS_FILE.  Register
  pass_started(self) and either pass_done(self) or
  pass_continued(self), depending on whether the last pass was carried
  out.