Use a new format for the unsynced database of CVSItems.
This database was write-once, and read in almost-sequential order.
The only non-sequential reads were read-accesses to other CVSItems for
the current file (when resyncing).  Therefore using a DB as storage
was overkill.

Now this data is stored in a simple file containing a sequence of
pickles.  The first pickle is a primer, and each subsequent pickle is
a list of CVSItems pertaining to a single file.  The pickles are read
one by one, and all of the CVSItems relating to the current file are
kept in memory until processing of that file is done.  This also
removes the need for 'cvs2svn-items-all.txt'.

This change cuts the size of 'cvs2svn-cvs-items.pck' (the new name for
'cvs2svn-cvs-items.db') in half and cuts the combined time for
CollectRevsPass and ResyncRevsPass significantly, too.

* cvs2svn_lib/cvs_item_database.py (NewCVSItemStore, OldCVSItemStore):
  New classes.

* cvs2svn_lib/config.py (CVS_ITEMS_ALL_DATAFILE): Remove constant
  (file is not needed anymore).

  (CVS_ITEMS_STORE): New constant (replaces CVS_ITEMS_DB).

* cvs2svn_lib/collect_data.py (CollectData._cvs_item_store): New
  member (replaces _cvs_items_db and _all_revs).

  (CollectData.add_cvs_item): Store item to _cvs_item_store instead of
  database.

  (CollectData.flush): New method (replaces write_symbol_stats()).
  Close _cvs_item_store as well.

* cvs2svn_lib/passes.py (CollectRevsPass.register_artifacts,
  ResyncRevsPass.register_artifacts): Register CVS_ITEMS_STORE instead
  of CVS_ITEMS_DB and CVS_ITEMS_ALL_DATAFILE.

  (CollectRevsPass.run): Call CollectData.flush() instead of
  CollectData.write_symbol_stats().

  (ResyncRevsPass.run): Use OldCVSItemStore instead of
  CVSItemDatabase.  Read via iterator rather than by id.

* design-notes.txt: Update documentation accordingly.