Eliminate the resynchronization of CVSItems.
This change makes issue #100 goes away.  But the fix is accidental.
The real cause of issue #100 is a missing dependency between vendor
branch and trunk that still has to be added.

Soon, ResyncRevsPass will pass.  But for now I leave it in just to
copy the cvs items to the random-access database.

* cvs2svn_lib/passes.py (CollectRevsPass.register_artifacts): Don't
  register RESYNC_DATAFILE.

  (ResyncRevsPass.register_artifacts): Don't register RESYNC_DATAFILE.

  (ResyncRevsPass.run): Don't resynchronize items; just copy them to
  the new database.

* cvs2svn_lib/collect_data.py
  (_FileDataCollector._process_revision_data): Don't write resync data
  file.

  (CollectData.__init__): Don't open resync data file.

* cvs2svn_lib/config.py (RESYNC_DATAFILE): Remove constant.

* cvs2svn_lib/cvs_revision_resynchronizer.py: Delete module.

* run-tests.py (test_list): Remove the XFail annotation on issue_100,
  since the test passes now.