Include an artificial dependency to make revision-only dependencies work.
Ideally, the first revision on a branch should only depend on the
CVSBranch, and the CVSBranch on the trunk revision from which the
branch sprouted.  But for now we are only considering the dependency
graph that includes CVSRevisions, so the exclusion of the intermediate
CVSBranch item causes this dependency to get lost.  So add an explicit
"helper" dependency directly between the trunk revision and the first
branch revision.

This causes issues #99 and #106 to pass, but I haven't analyzed
whether the fixes are accidental.  (One dependency is still known to
be missing.)

* cvs2svn_lib/cvs_item.py (CVSRevision.branch_commit_ids): New member,
  recording the first commits on any branches that spring from this
  revision.

  (CVSRevision.__init__, CVSRevision.__getstate__,
  CVSRevision.__setstate__): Initialize, save, and restore new member.

  (CVSRevision.get_pred_ids, CVSRevision.get_succ_ids): Add the helper
  dependency described above.

* cvs2svn_lib/collect_data.py (_RevisionData.branches_revs_data): New
  member.

  (_FileDataCollector._resolve_dependencies): Set branches_revs_data.

  (_FileDataCollector._process_revision_data): Pass branch_commit_ids
  parameter to CVSRevision constructor.

* cvs2svn_lib/passes.py (FilterSymbolsPass.filter_excluded_symbols):
  Delete the first commit on an excluded branch from the
  branch_commit_ids of the CVSRevision from which the branch springs.

* run-tests.py (test_list): Remove XFail annotation from issue_99 and
  issue_106.