Break up loops involving symbol dependencies.
Break up loops in the commitset dependency graph that involve symbol
changesets.  These loops are broken by breaking up branch-creation
changesets.  After this step, the symbol changesets are discarded.
But their temporary presence has an effect on the conversion because
they cause revision changesets to be processed in a different order.

The program seems to run with this change, but many tests fail because
of this algorithm's preference for changing commit times over breaking
up branch changesets.  (I wrote an email to the dev mailing list about
this problem.)  The algorithm will have to be improved before it can
be used in trunk.

* cvs2svn_lib/passes.py
  (BreakCVSRevisionChangesetLoopsPass.register_artifacts): Change
  artifact order to be consistent with other passes.

  (BreakCVSSymbolChangesetLoopsPass): New pass class.

  (TopologicalSortPass.run): Use CHANGESETS_SYMBROKEN_DB and
  CVS_ITEM_TO_CHANGESET_SYMBROKEN instead of
  config.CHANGESETS_REVBROKEN_DB and CVS_ITEM_TO_CHANGESET_REVBROKEN.
  Do topological sort including symbol changesets, but only outpu
  revision changesets.

  (TopologicalSortPass.register_artifacts,
  CreateDatabasesPass.register_artifacts,
  CreateDatabasesPass.get_changesets,
  CreateRevsPass.register_artifacts, CreateRevsPass.get_changesets):
  Use CHANGESETS_SYMBROKEN_DB and CVS_ITEM_TO_CHANGESET_SYMBROKEN
  instead of config.CHANGESETS_REVBROKEN_DB and
  CVS_ITEM_TO_CHANGESET_REVBROKEN.

* cvs2svn (pass_manager): Add BreakCVSSymbolChangesetLoopsPass.

* cvs2svn_lib/config.py (CVS_ITEM_TO_CHANGESET_SYMBROKEN,
  CHANGESETS_SYMBROKEN_DB): New constants.

* doc/design-notes.txt, doc/symbol-notes.txt: Document new pass.