Remove the workaround for an old Python pickling bug.
In Python 2.2, there was a bug preventing Python from finding custom
pickling methods __setstate__() and __getstate__() that were inherited
from a base class if the derived class defined __slots__.  The
workaround was to set those methods explicitly in each derived class.
Now that we require Python 2.4, this workaround is not needed anymore.

* cvs2svn_lib/cvs_item.py (CVSRevisionModification.__setstate__,
  CVSRevisionModification.__getstate, CVSRevisionAdd.__setstate__,
  CVSRevisionAdd.__getstate__, CVSRevisionChange.__setstate__,
  CVSRevisionChange.__getstate__, CVSRevisionAbsent.__setstate__,
  CVSRevisionAbsent.__getstate__, CVSRevisionDelete.__setstate__,
  CVSRevisionDelete.__getstate__, CVSRevisionNoop.__setstate__,
  CVSRevisionNoop.__getstate__, CVSBranchNoop.__setstate__,
  CVSBranchNoop.__getstate__, CVSTagNoop.__setstate__,
  CVSTagNoop.__getstate__): Remove redundant method definitions.