Remove no-op revisions
Patch by: Jon Foster <jon.foster@cabot.co.uk>

Note from me: It is debatable whether this change is useful enough (or
reviewed enough) to be added to cvs2svn.  Nevertheless, I am hereby
committing it on a branch for posterity; perhaps somebody would like
to do something with it.  See the following thread for some
discussion:

http://cvs2svn.tigris.org/ds/viewMessage.do?dsForumId=1667&dsMessageId=2455486

From author's email:

With CVS, you can use "cvs commit -f" to commit a file that hasn't
actually changed.  This generates a new revision in the repository,
with an empty diff.  This patch adds a --remove-noop-revisions flag to
cvs2svn, which causes (most of)* these no-op CVS commits to be
ignored.

If it detects a no-op commit, it will delete the relevant revision and
graft any tags/branches onto the previous revision.

There is a TODO in this patch: I'm not sure if I need to sort the
tags/branches into chronological order when grafting them.  Does the
order matter?

There is a simple test-case included.

* This code works by looking to see if the RCS patch between two
adjacent revisions is empty.  As such, it can't detect the case where
you do a "cvs import", possibly followed by more imports on the vendor
branch, then do a no-op commit that creates rev 1.2.  This is because
rev 1.1.1.3 (or whatever the last vendor-branch revision is) and rev
1.2 aren't adjacent revisions in the RCS file.  Detecting that would
require a lot more work and would slow down cvs2svn - it would likely
have to generate fulltext for rev 1.2 and the last vendor-branch
revision, to see if they're the same.  So I don't think it's worth it.