Change options file to show how to use ExternalBlobGenerator.
Add an example in the options file (commented out) showing how to use
the ExternalBlobGenerator with cvs2git.  This method is vastly faster
than the old method of generating fulltext blobs using CVS -- by a
factor of 20 in the case of the Gentoo repository (which is admittedly
an unusual case).

In the future this code might be made the default, but first it needs
some more testing.