Allow the svn:executable property to affect the git filemode
This allows a file to be marked executable in Git using the full
property-setting machinery, for example via `AutoPropsPropertySetter`.

We only need to check for the presence of the `svn:executable` flag
because Git already uses the `ExecutablePropertySetter`, which sets
the `svn:executable` property based on whether the CVS file had its
executable bit set.

Patch by: Mark Rotteveel <mark@lawinegevaar.nl>